[MOD] LottoGenerator 클래스를 AutoLotto 클래스와 LottoGenerator 클래스로 분리
LottoGenerator클래스를 구현하다보니 인스턴스 변수가 많아지고 AutoLotto의 객체의 필요성을 느껴서 AutoLotto에서 GenerateAuttoLotto, GenerateRandomNumber, isOverlapLottoNumber 를 구현하였습니다.
