+++ src/main/java/domain/LottoGenerator.java
@@ -4,55 +4,27 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class LottoGenerator {
-    private static final int PRIZE_OF_LOTTO = 1000;
-    private static final int MAX_LOTTO_NUMBER = 45;
-    private static final int MIN_LOTTO_NUMBER = 1;
-    private static final int LOTTO_NUMBER_COUNT = 6;
-
     private static List<Lotto> lottoList;
-    private List<Integer> lottoNumbers;
 
     LottoGenerator() {
         lottoList = new ArrayList<>();
-        lottoNumbers = new ArrayList<>();
     }
 
-    public void GenerateLottos(int purchaseAmount) {
-        int lottoCount = purchaseAmount / PRIZE_OF_LOTTO;
+    public void GenerateLottos(int lottoCount) {
+
         for (int i = 0; i < lottoCount; i++) {
             GenerateLotto();
         }
     }
 
     private void GenerateLotto() {
-        for (int lottoNumberCount = 1; lottoNumberCount <= LOTTO_NUMBER_COUNT; lottoNumberCount++) {
-            lottoNumbers.add(GenerateRandomNumber());
-        }
+        AutoLotto autoLotto = new AutoLotto();
 
-        lottoList.add(new Lotto(lottoNumbers));
-        lottoNumbers.clear();
-
-    }
-
-    private int GenerateRandomNumber() {
-        boolean isNotValidLottoNumber = true;
-        int candidateLottoNumber = 0;
-
-        while (isNotValidLottoNumber) {
-            candidateLottoNumber = (int) ((Math.random() * MAX_LOTTO_NUMBER) + MIN_LOTTO_NUMBER);
-            isNotValidLottoNumber = isNotOverlapLottoNumber(candidateLottoNumber);
-        }
-
-        return candidateLottoNumber;
-    }
-
-    private boolean isNotOverlapLottoNumber(int candidateLottoNumber) {
-        return lottoNumbers.contains(candidateLottoNumber) ? false : true;
+        lottoList.add(new Lotto(autoLotto.GenerateAuttoLotto()));
     }
 
     public static List<Lotto> getLottoList() {
         return lottoList;
     }
-
 
 }
